+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
 
+<<<<<<< HEAD
+- feat(text_extraction/): Se agregó módulo text_extraction
+=======
 - build(requirements/): Se agregaron los archivos de requirements/ para instalación con pip
+>>>>>>> dfa3bcf4e7b1850b10f9b5ed0c580d179f2c6ac2
 - Initial commit